Remove future shadow.
