+++ examples/example-core/example-raw-jdbc/src/main/java/org/apache/shardingsphere/example/core/jdbc/repository/ShadowUserRepositoryImpl.java
@@ -18,7 +18,7 @@
 package org.apache.shardingsphere.example.core.jdbc.repository;
 
 import org.apache.shardingsphere.example.core.api.entity.ShadowUser;
-import org.apache.shardingsphere.example.core.api.repository.ShadowUserRepository;
+import org.apache.shardingsphere.example.core.api.repository.CommonRepository;
 
 import javax.sql.DataSource;
 import java.sql.Connection;
@@ -29,7 +29,7 @@ import java.sql.Statement;
 import java.util.LinkedList;
 import java.util.List;
 
-public final class ShadowUserRepositoryImpl implements ShadowUserRepository {
+public final class ShadowUserRepositoryImpl implements CommonRepository<ShadowUser, Long> {
     
     private final DataSource dataSource;
     
@@ -40,7 +40,7 @@ public final class ShadowUserRepositoryImpl implements ShadowUserRepository {
     @Override
     public void createTableIfNotExists() throws SQLException {
         String sql = "CREATE TABLE IF NOT EXISTS t_user "
-                + "(user_id INT NOT NULL AUTO_INCREMENT, user_name VARCHAR(200), pwd VARCHAR(200), PRIMARY KEY (user_id))";
+                + "(user_id INT NOT NULL AUTO_INCREMENT, user_type INT NOT NULL, user_name VARCHAR(200), pwd VARCHAR(200), PRIMARY KEY (user_id))";
         try (Connection connection = dataSource.getConnection();
              Statement statement = connection.createStatement()) {
             statement.executeUpdate(sql);
@@ -67,13 +67,13 @@ public final class ShadowUserRepositoryImpl implements ShadowUserRepository {
     
     @Override
     public Long insert(final ShadowUser entity) throws SQLException {
-        String sql = "INSERT INTO t_user (user_id, user_name, pwd,shadow) VALUES (?, ?, ?,?)";
+        String sql = "INSERT INTO t_user (user_id, user_type, user_name, pwd) VALUES (?, ?, ?, ?)";
         try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
             preparedStatement.setInt(1, entity.getUserId());
-            preparedStatement.setString(2, entity.getUserName());
-            preparedStatement.setString(3, entity.getPwd());
-            preparedStatement.setBoolean(4, entity.isShadow());
+            preparedStatement.setInt(2, entity.getUserType());
+            preparedStatement.setString(3, entity.getUserName());
+            preparedStatement.setString(4, entity.getPwd());
             preparedStatement.executeUpdate();
         }
         return (long) entity.getUserId();
@@ -81,38 +81,38 @@ public final class ShadowUserRepositoryImpl implements ShadowUserRepository {
     
     @Override
     public void delete(final Long id) throws SQLException {
-        String sql = "DELETE FROM t_user WHERE user_id=? and shadow=?";
-        deleteUser(sql, id, true);
-        deleteUser(sql, id, false);
+        String sql = "DELETE FROM t_user WHERE user_id = ? and shadow= ?";
+        deleteUser(sql, id, (int) (id % 2));
+        deleteUser(sql, id, (int) (id % 2));
     }
     
-    private void deleteUser(final String sql, final Long id, final boolean isShadow) throws SQLException {
+    private void deleteUser(final String sql, final Long id, final int userType) throws SQLException {
         try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
             preparedStatement.setLong(1, id);
-            preparedStatement.setBoolean(2, isShadow);
+            preparedStatement.setInt(2, userType);
             preparedStatement.executeUpdate();
         }
     }
     
     @Override
     public List<ShadowUser> selectAll() throws SQLException {
-        List<ShadowUser> users;
-        String sql = "SELECT * FROM t_user where shadow=?";
-        users = getUsers(sql, false);
-        users.addAll(getUsers(sql, true));
+        String sql = "SELECT * FROM t_user where user_type = ?";
+        List<ShadowUser> users = getUsers(sql, 1);
+        users.addAll(getUsers(sql, 0));
         return users;
     }
     
-    private List<ShadowUser> getUsers(final String sql, final boolean isShadow) throws SQLException {
+    private List<ShadowUser> getUsers(final String sql, final int userType) throws SQLException {
         List<ShadowUser> result = new LinkedList<>();
         try (Connection connection = dataSource.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(sql)) {
-            preparedStatement.setBoolean(1, isShadow);
+            preparedStatement.setInt(1, userType);
             ResultSet resultSet = preparedStatement.executeQuery();
             while (resultSet.next()) {
                 ShadowUser user = new ShadowUser();
                 user.setUserId(resultSet.getInt("user_id"));
+                user.setUserType(resultSet.getInt("user_type"));
                 user.setUserName(resultSet.getString("user_name"));
                 user.setPwd(resultSet.getString("pwd"));
                 result.add(user);
